+++ packages/contracts/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added hash collision check to `createVersion` in `PluginRepo`.
 - Created the `IMultisig` interface.
 
 ### Changed